style: Move varint encoding to separate module
While the `encoding` module is undocumented and not stable, there are known users. Leave a alias in place to prevent breaking them.
